refactor: move FlowDebugOutput and CommandDebugMetadata to maestro-orchestra
Data types for debug output now live in maestro.orchestra.debug so the
cloud worker can consume them. TestDebugReporter.kt in CLI retains only
the CLI-specific lifecycle (install, saveSuggestions, AI output types,
etc.). No behavior change; all CLI call sites updated.
