+++ maestro-cli/src/main/java/maestro/cli/runner/MaestroCommandRunner.kt
@@ -23,9 +23,7 @@ import maestro.Maestro
 import maestro.MaestroException
 import maestro.device.Device
 import maestro.cli.report.SingleScreenFlowAIOutput
-import maestro.cli.report.CommandDebugMetadata
 import maestro.cli.report.FlowAIOutput
-import maestro.cli.report.FlowDebugOutput
 import maestro.cli.runner.resultview.ResultView
 import maestro.cli.runner.resultview.UiState
 import maestro.cli.util.PrintUtils
@@ -33,7 +31,9 @@ import maestro.orchestra.ApplyConfigurationCommand
 import maestro.orchestra.CompositeCommand
 import maestro.orchestra.MaestroCommand
 import maestro.orchestra.Orchestra
+import maestro.orchestra.debug.CommandDebugMetadata
 import maestro.orchestra.debug.CommandStatus
+import maestro.orchestra.debug.FlowDebugOutput
 
 import maestro.orchestra.yaml.YamlCommandReader
 import maestro.utils.CliInsights
